Add support for Data Sources beyond files.
__Overview__

The current design of Data Sources in the SDK is veryWindows-centric in that
it assumes that everything is a file with an extension.

Three scenarios are driving this change:
- Plugins that need to process Directories rather than Files
- Plugins that need to process Files that do not have extensions
- Plugins that need to process something other than a File or Directory

This change generalizes away from file to enable support of arbitrary Data Sources,
and in keeping with the existing API, adds two (2) new attributes for
authors to use on their Custom Data Sources:
- ExtensionlessFileDataSourceAttribute
- DirectoryDataSourceAttribute

The _ExtensionlessFileDataSourceAttribute_ is used to denote that Files that do
not have extensions may be routed to the Custom Data Source.

The _DirectoryDataSourceAttribute_ is used to denote that Directories may be routed
to the Custom Data Source.

In order to support these attributes, as well as any arbitrary Data Sources, we
have generalized the ability for Custom Data Sources to state that they support
a Data Source. _IsFileSupported has been changed to IsDataSourceSupported_ and now
takes an _IDataSource_ rather than a string. This allows for any _IDataSource_
implementation to be interrogated, even user defined implemetations.

The existing attributes for Files and Directories are now used simply to enhance
the performance of assigning Data Sources to their appropriate Custom Data Sources
for processing. They are also used by UI components in their Open dialogs. Unrecognized
implementations of _IDataSource_ will simply be routed to all Custom Data Sources and
have their _IsDataSourceSupported_ method called.

The _CustomDataSourceReference_ class has been updated to work in terms of _IDataSource_s
as well, and the myriad of extension methods around this class and the _ICustomDataSource_
interface have been refactored to better reflect the now generalized world. This class
has also been made abstract in order to allow for easier testing of components that utilize
_CustomDataSourceReference_s.

Finally, the _Engine_ has been updated to operate in terms of _IDataSource_ rather than
string paths. The old _AddFile_ methods have been converted to extension methods for
convenience.

__Details__

- Custom Data Source interfaces and base classes updated to use the more general
  IDataSource
- New Data Source Attributes added to enable additional common scenarios.
- Support for arbitrary IDataSources.
- DataSourceResolver introduced to eliminate duplication of assigning Data Sources
  to the Custom Data Sources that can process said Data Sources.
- Engine updated to operate on IDataSource rather than strings.
- CustomDataSourceReference updated to allow for better mocking in tests.
- Additional Tests added
- Existing tests updated
- NBGV updated due to breaks on build machines.

__Breaking Changes__

This change introduces a number of breaking changes. They are detailed below.

- Microsoft.Performance.SDK.Runtime.CustomDataSourceExtensions
	- Removed method TryGetFileExtension
	- Removed method TryGetFileDescription
	- Removed method Supports

- Microsoft.Performance.SDK.Runtime.CustomDataSourceExtensions
  - Removed field EqualityComparer

- Microsoft.Performance.SDK.Runtime.CustomDataSourceReferenceExtensions
	- Renamed TryGetFileExtension to TryGetFileExtensions
	- Renamed TryGetCanonicalFileExtension to TryGetCanonicalFileExtensions
	- TryGetFileDescription now has a required parameter

- Microsoft.Performance.SDK.Processing.CustomDataSourceBase
	- Removed method IsFileSupported
	- Removed method IsFileSupportedCore
	- Added method IsDataSourceSupportedCore

- Microsoft.Performance.SDK.Processing.ICustomDataSource
	- Removed method IsFileSupported
	- Added method IsDataSourceSupported

- Microsoft.Performance.SDK.ReadOnlyHashSet<T>
	- Now implements IReadOnlyCollection<T>

- Microsoft.Performance.Toolkit.Engine.Engine
	- Renamed FreeFilesToProcess to FreeDataSourcesToProcess
	- Renamed FilesToProcess to DataSourcesToProcess

- Renamed Microsoft.Performance.Toolkit.EngineUnsupportedDataSourceException to UnsupportedCustomDataSourceException

- Renamed Microsoft.Performance.Toolkit.EngineUnsupportedFileException to UnsupportedDataSourceException