Enforce SDK Version Compatibility.
We introduce the concept of checking the compatability of a Plugin
with the version of the SDK being hosted.

A Plugin is considered to be using a Version of the SDK compatible with
the Host if the following are true:
1) The Plugin is using SDK Version greater than or equal to the lowest supported version (currently 0.109.0)
2) The Major Version of the Plugin's referenced SDK is equal to the  Major version of the hosted SDK.
3) If the Major Version is 1 or greater, then the Minor version of the plugin must be less than or equal
to the Minor version of the SDK. If the Major Version is zero, then a simple range check is used to make
sure the Plugin version is greater than or equal to the lowest supported and less than or equal to the
current SDK version.

If either of these checks fails, then the Plugin is rejected as incompatible
with the host.

_IMPLEMENTATION_

We added a class to do the version checks. This class is used deep in the
Discovery mechanism to do the check as close to loading as possible. This
gives us the opportunity to pinpoint the incompatible to the assembly.

We implemented a pass through of the ErrorInfo class in order to propogate
all errors during Discovery up to the PluginLoader. Before, we had a simple
Boolean, and so the nature of errors was getting lost. In order to tell
the user what all has gone wrong, we MUST propogate error information.

All in all, this change checks compatability, and provides for propogating
all errors up the call stack back to the loader so that failures can be
reported conveniently and accurately by the host to the end user.

__Breaking Changes__

This change introduces a number of breaking changes. They are detailed below.

- Microsoft.Performance.SDK.Runtime.NetCoreapp.PluginsLoader
  - TryLoadPlugins(IEnumerable<string> directories, out IList<string> failed) changed to TryLoadPlugins(IEnumerable<string> directories, out IDictionary<string, ErrorInfo> failed)
  - Task<(bool, IList<string>)> TryLoadPluginsAsync(IEnumerable<string> directories) changed ti Task<(bool, IDictionary<string, ErrorInfo>)> TryLoadPluginsAsync(IEnumerable<string> directories)

- Microsoft.Performance.SDK.Runtime.Discovery.AssemblyExtensionDiscovery
  - All ProcessAssemblies methods now have an 'out ErrorInfo' parameter