Generalize Data Source Attributes.
This change extends the DataSourceAttribute base attribute
to define the Type of Data Source accepted by the decorated
Custom Data Source.

Users may implement a preliminary filtering method in order to
shrink the space of valid Data Sources (for example, the File Data Source
shrinks the acceptable space to files with a certain extension.)

Data Sources declared by a Custom Data Source will be routed to said
Custom Data Source for determining if the Data Source is supported.

Data Sources that are unrecognized do not get routed to any Custom Data
Sources.

__Breaking Changes__

This change introduces a number of breaking changes. They are detailed below.

- Microsoft.Performance.SDK.Runtime.DataSource
  - moved to Microsoft.Performance.SDK.Processing.DataSource
  - Microsoft.Performance.SDK.Runtime.DataSourceExtensions
  - moved to Microsoft.Performance.SDK.Processing.DataSourceExtensions
- Microsoft.Performance.SDK.Runtime.FileDataSource
  - moved to Microsoft.Performance.SDK.Processing.FileDataSource
- Microsoft.Performance.SDK.Processing.DataSourceAttribute
  - a new parameter has been added to the constructors