Do not enumerate types from non-extension dlls.
Not all DLLs reference the SDK. If a DLL does not reference the SDK, then
there is no way that said DLL can contain a plugin or extension. This change
will cause the extension discovery to skip over any DLLs that do not reference
the SDK when doing Type enumeration.

After validating that the DLL is acceptable according to the validator, it is
loaded and passed to ProcessAssembly. Here, we add a check for an SDK reference
and return early if the reference is not present.

Tests have been added for this scenario.