Update CustomDataSourceReference Supports to filter correctly.
CustomDataSourceReference was blindly passing datasources directly to the underlying
CustomDataSource without checking if it was even remotely plausible that the CDS could
accept the data source.

The supports method has been updated to interrogate the data source attributes before
asking the Custom Data Source.

The CustomDataSOurceReference class had to be refactored to make this testable; all
tests were updated to use the CustomDataSourceReference class directly and still pass without
changing behavior. This also has the side effect of making the CustomDataSourceReference
class easier to understand.