PluginsSystem: consolidate plugin information into PluginInfo class
AvailablePluginInfo, InstalledPluginInfo, and PluginMetadata was
declaring identical properties, leading to duplicated code and sources
of truth. This PR consolidates these properties into a PluginInfo class
that is exposed on AvailablePluginInfo and InstalledPluginInfo.

Instead of also exposing this class on PluginMetadata, this PR also adds
a new file to plugin packages that contains a serialization of a
plugin's PluginInfo.